requirements.txt
openai>=1.30.0
pymongo>=4.7.0
voyageai>=0.3.0
mcp[cli]>=1.2.0
python-dotenv>=1.0.0
